AdvPy25/Lista07/requirements.txt
aiohttp
beautifulsoup4
pillow
rich
rich-pixels